Add Component For Modals

+++ src/page/Home.js
@@ -1,9 +1,10 @@
-import React, { createContext, useContext, useEffect, useState } from "react";
+import React, { useEffect, useState } from "react";
 import Header from "../component/Header";
 import DetailsModal from "../component/DetailsModal";
 import axios from "axios";
 import { useFilter } from "../component/Filterdata";
 import Pagination from "../component/Pagination";
+import AddNewDataModal from "../component/Modals/AddNewDataModal";
 
 export default function Home() {
 
@@ -28,7 +29,7 @@ export default function Home() {
   const [nameicon, setNameicon] = useState("down");
   const [departmenticon, setDepartmenticon] = useState("down");
   const [idicon, setIdicon] = useState("down");
-  const { department, updateDepartment, updateList, filter, icon } = useFilter();
+  const { department, updateDepartment, updateList, filter } = useFilter();
 
   const [itemsperpage, setItemsperpage] = useState(10);
   const [currentpage, setCurrentpage] = useState(1);
@@ -37,8 +38,8 @@ export default function Home() {
     pages.push(i);
   }
   const indexOfLastItem = currentpage * itemsperpage;
-  const indexOfFirestItem = indexOfLastItem - itemsperpage;
-  
+  const indexOfFirstItem = indexOfLastItem - itemsperpage;
+
   let UI = 0;
   let react = 0;
   let Python = 0;
@@ -55,6 +56,7 @@ export default function Home() {
 
   const handleShow = (data) => {
     setShowModal(true);
+    setModaltitle("More Details")
     setEmail(data.email);
     setPhone(data.phone);
     setExp(data.experience);
@@ -82,7 +84,6 @@ export default function Home() {
       setReject(false)
 
     } catch (err) {
-      console.log(err)
       setLoad(false)
       setReject(true)
     }
@@ -105,6 +106,7 @@ export default function Home() {
   }, [department]);
 
   useEffect(() => {
+
     const FilteredDepartment = Newfdata.filter((val) =>
       val.department.toLowerCase().includes(Search.toLowerCase())
     )
@@ -115,23 +117,26 @@ export default function Home() {
       val.id === Search
     )
 
-    // let thispage = currentpage;
     const FilteredSet = new Set([...FilteredDepartment, ...FilteredName, ...FilteredId])
     const FilteredResults = [...FilteredSet]
     FilteredResults.length > 0 ? setFdata(FilteredResults) : setFdata([]);
 
-    // FilteredResults.length > 0 ? setCurrentpage(1) : setCurrentpage(thispage);
-    // setMaxPageNumberLimit(5)
-    // setMinPageNumberLimit(0)
+    setCurrentpage(1);
   }, [Search, Newfdata])
 
   const DeleteItem = async (id) => {
-    try {
-      await axios.delete(`http://192.168.11.198:8000/employees/${id}`)
-      FetchData();
+
+    if (window.confirm("Are You Sure To Delete This Record?")) {
+
+      try {
+        await axios.delete(`http://192.168.11.198:8000/employees/${id}`)
+        FetchData();
+      }
+      catch (err) {
+      }
     }
-    catch (err) {
-      console.log(err)
+    else {
+      return false;
     }
   }
 
@@ -163,56 +168,59 @@ export default function Home() {
 
   const AddNewData = async () => {
 
-    if (NewData.name == "" || NewData.department == "" || NewData.experience == "" || NewData.email == "" || NewData.salary == "" || NewData.phone == "") {
-      alert("Please fill all the fields")
-    }
-    else if (NewData.department !== "REACT" && NewData.department !== "PYTHON" && NewData.department !== "DB" && NewData.department !== "UI/UX") {
-      alert("Department Not Allowed!");
-    }
-    else if (!validateEmail(NewData.email)) {
-      alert("Enter a valid email address");
-      return false;
-    }
-    else if ((NewData.phone.toLowerCase() >= 'a' && NewData.phone.toLowerCase() <= 'z') || (NewData.phone.length != 10)) {
-      alert("Enter Valid Phone Number");
-    }
-    else if ((NewData.salary.toLowerCase() >= 'a' && NewData.salary.toLowerCase() <= 'z')) {
-      alert("Enter Valid Salary");
-    }
+    switch (true) {
+      case (NewData.name == "" || NewData.department == "" || NewData.experience == "" || NewData.email == "" || NewData.salary == "" || NewData.phone == ""):
+        alert("Please fill all the fields")
+        break;
 
-    else {
-      if (AddUpdatebtn === "Add") {
-        try {
-          await axios.post('http://192.168.11.198:8000/employees', NewData)
-          setAddmodal(false)
-          setModaltitle("Update Data")
-          FetchData();
-        }
-        catch (err) {
-          console.log(err)
-          alert('Failed To Add Data');
-        }
-      }
+      case (NewData.department !== "REACT" && NewData.department !== "PYTHON" && NewData.department !== "DB" && NewData.department !== "UI/UX"):
+        alert("Department Not Allowed!");
+        break;
 
-      else {
-        NewData.id = Id;
-        try {
-          await axios.put(`http://192.168.11.198:8000/employees/${Id}`, NewData)
-          setName("")
-          setDepartment("")
-          setExp("")
-          setEmail("")
-          setSalary("")
-          setPhone("")
-          setModaltitle("Add Data");
-          setAddmodal(false)
-          FetchData();
+      case (!validateEmail(NewData.email)):
+        alert("Enter a valid email address");
+        return false;
+
+      case ((NewData.phone.toLowerCase() >= 'a' && NewData.phone.toLowerCase() <= 'z') || (NewData.phone.length != 10)):
+        alert("Enter Valid Phone Number");
+        break;
+
+      case ((NewData.salary.toLowerCase() >= 'a' && NewData.salary.toLowerCase() <= 'z')):
+        alert("Enter Valid Salary");
+        break;
+
+      default:
+        if (AddUpdatebtn === "Add") {
+          try {
+            await axios.post('http://192.168.11.198:8000/employees', NewData)
+            setAddmodal(false)
+            setModaltitle("Update Data")
+            FetchData();
+          }
+          catch (err) {
+            alert('Failed To Add Data');
+          }
         }
-        catch (err) {
-          console.log(err)
-          alert('Failed To Update Data');
+
+        else {
+          NewData.id = Id;
+          try {
+            await axios.put(`http://192.168.11.198:8000/employees/${Id}`, NewData)
+            setName("")
+            setDepartment("")
+            setExp("")
+            setEmail("")
+            setSalary("")
+            setPhone("")
+            setModaltitle("Add Data");
+            setAddmodal(false)
+            FetchData();
+          }
+          catch (err) {
+            alert('Failed To Update Data');
+          }
         }
-      }
+        break;
     }
   }
 
@@ -242,35 +250,31 @@ export default function Home() {
     if (filter === "id") {
       if (idicon === "up") {
         setFdata([...Newfdata].sort((a, b) => b.id - a.id))
-        console.log("true")
       }
       else {
         setFdata([...Newfdata].sort((a, b) => a.id - b.id));
-        console.log("false")
+
       }
     }
 
     else if (filter === "name") {
       if (nameicon === "up") {
-        setFdata([...Newfdata].sort((a, b) => a.name.toLowerCase().localeCompare(b.name.toLowerCase())))
-      }
-      else {
         setFdata([...Newfdata].sort((a, b) => b.name.toLowerCase().localeCompare(a.name.toLowerCase())))
       }
+      else {
+        setFdata([...Newfdata].sort((a, b) => a.name.toLowerCase().localeCompare(b.name.toLowerCase())))
+      }
     }
-
     else if (filter === "department") {
       if (departmenticon === "up") {
-        setFdata([...Newfdata].sort((a, b) => a.department.toLowerCase().localeCompare(b.department.toLowerCase())))
+        setFdata([...Newfdata].sort((a, b) => b.department.toLowerCase().localeCompare(a.department.toLowerCase())))
       }
       else {
-        setFdata([...Newfdata].sort((a, b) => b.department.toLowerCase().localeCompare(a.department.toLowerCase())))
-
+        setFdata([...Newfdata].sort((a, b) => a.department.toLowerCase().localeCompare(b.department.toLowerCase())))
       }
     }
   }, [Newfdata, filter, idicon, nameicon, departmenticon])
-
-
+  var hello = "hello";
   return (
     <>
       <Header />
@@ -318,13 +322,12 @@ export default function Home() {
                     </div>
 
                     <div className="container table_hight">
-                      <table className="table table-light table-bordered">
+                      <table className="table table-bordered">
                         <thead className="table-header">
                           <tr>
                             <th scope="col">Id<i className={`ms-2 fa-solid fa-angle-${idicon}`} onClick={() => { return handleListFilter("id"), idicon === "up" ? setIdicon("down") : setIdicon("up") }}></i></th>
                             <th scope="col">Name<i className={`ms-2 fa-solid fa-angle-${nameicon}`} onClick={() => { return handleListFilter("name"), nameicon === "up" ? setNameicon("down") : setNameicon("up") }}></i></th>
                             <th scope="col">Department<i className={`ms-2 fa-solid fa-angle-${departmenticon}`} onClick={() => { return handleListFilter("department"), departmenticon === "up" ? setDepartmenticon("down") : setDepartmenticon("up") }}></i></th>
-
                             <th scope="col">Action</th>
                           </tr>
                         </thead>
@@ -336,7 +339,7 @@ export default function Home() {
                                 <td colSpan="4">Search Item Not Found</td>
                               </tr>
                               :
-                              Fdata.slice(indexOfFirestItem, indexOfLastItem).map((val, index) => {
+                              Fdata.slice(indexOfFirstItem, indexOfLastItem).map((val, index) => {
                                 return (
                                   <>
                                     <tr key={index}>
@@ -352,9 +355,6 @@ export default function Home() {
 
                                       </td>
                                     </tr>
-
-
-
                                   </>
                                 )
                               })
@@ -364,77 +364,43 @@ export default function Home() {
 
                       {ShowModal && (
                         <>
-                          <div className="modal show" tabIndex="-1" role="dialog" style={{ display: 'block' }}>
-                            <div className="modal-dialog" role="document">
-                              <div className="modal-content">
-
-                                <div className="modal-header">
-                                  <h5 className="modal-title">More Details</h5>
-                                  <button type="button" className="close" aria-label="Close" onClick={handleClose}>
-                                    <span aria-hidden="true">&times;</span>
-                                  </button>
-                                </div>
-
-                                <div className="modal-body text-start">
-                                  <h5>Email</h5>
-                                  <p>{Email}</p>
-                                  <h5>Phone</h5>
-                                  <p>{Phone}</p>
-                                  <h5>Experience</h5>
-                                  <p>{Exp} Years</p>
-                                  <h5>Salary</h5>
-                                  <p>{Salary}</p>
-                                </div>
-
-                                <div className="modal-footer">
-                                  <button type="button" className="btn btn-secondary" onClick={handleClose}>
-                                    Close
-                                  </button>
-                                </div>
-
-                              </div>
-                            </div>
-                          </div>
-                          <div className="modal-backdrop show"></div>
+                          <DetailsModal
+                            Email={Email}
+                            Phone={Phone}
+                            Exp={Exp}
+                            Salary={Salary}
+                            handleClose={handleClose}
+                            Modaltitle={Modaltitle}
+                          />
                         </>
                       )}
                       {Addmodal && (
                         <>
-                          <div className='modal show' tabIndex="-1" role="dialog" style={{ display: 'block' }}>
-                            <div className="modal-dialog" role="document">
-                              <div className="modal-content">
-                                <div className="modal-header">
-                                  <h5 className="modal-title">{Modaltitle}</h5>
-                                  <button type="button" className="close" onClick={handleClose} aria-label="Close">
-                                    <span aria-hidden="true">&times;</span>
-                                  </button>
-                                </div>
-                                <div className="modal-body text-start">
-                                  Name:<input type="text" className="form-control" value={Name} onChange={(evt) => { setName(evt.target.value) }} />
-                                  Department:<input type="text" className="form-control" value={Department} onChange={(evt) => setDepartment(evt.target.value)} />
-                                  Experience:<input type="text" className="form-control" value={Exp} onChange={(evt) => { setExp(evt.target.value) }} />
-                                  Email:<input type="email" className="form-control" value={Email} onChange={(evt) => { setEmail(evt.target.value) }} />
-                                  Phone:<input type="tel" className="form-control" value={Phone} onChange={(evt) => { setPhone(evt.target.value) }} />
-                                  salary:<input type="text" className="form-control" value={Salary} onChange={(evt) => { setSalary(evt.target.value) }} />
-                                </div>
-                                <div className="modal-footer">
-                                  <button className="btn btn-success" onClick={AddNewData}>{AddUpdatebtn}</button>
-                                  <button type="button" className="btn btn-secondary" onClick={handleClose}>
-                                    Close
-                                  </button>
-                                </div>
-                              </div>
-                            </div>
-                          </div>
-                          <div className="modal-backdrop show"></div>
+                          <DetailsModal
+                            Name={Name}
+                            setName={setName}
+                            Department={Department}
+                            setDepartment={setDepartment}
+                            Exp={Exp}
+                            setExp={setExp}
+                            Email={Email}
+                            setEmail={setEmail}
+                            Phone={Phone}
+                            setPhone={setPhone}
+                            Salary={Salary}
+                            setSalary={setSalary}
+                            handleClose={handleClose}
+                            AddUpdatebtn={AddUpdatebtn}
+                            AddNewData={AddNewData}
+                            Modaltitle={Modaltitle}
+                          />
                         </>
                       )}
 
-                    </div>
-                    <Pagination currentpage={currentpage} setCurrentpage={setCurrentpage} pages={pages}/>
-                    <DetailsModal />
-                  </div>
-                </div>
+                    </div >
+                    <Pagination currentpage={currentpage} setCurrentpage={setCurrentpage} pages={pages} />
+                  </div >
+                </div >
             }
           </>
       }
